Add rollForward policy to global.json
Without a rollForwardPolicy the developer has to have the EXACT version specified in the global.json.
Especially if you are not in a devcontainer, I think this is not that easy and not necessary.

Therefore I added the rollForward attribute to global.json, which also allows to use newer versions of the SDK instead of only the exact version. latestFeature is just a guess what could work nicely and could be changed.